Noddy implementation of preferred alias, needs tests adding
